{"types":"C","runs":131}

+++ midTPracProblem16.java
@@ -9,7 +9,7 @@ public class midTPracProblem16 extends ConsoleProgram{
 
 	
 	
-	private int findNthMember(double n, double a, double q){
+	private double findNthMember(double n, double a, double q){
 		n = n - 1;
 		double b = Math.pow(q, n);
 		double nthMember = a * b;
